goToDetailPage 로직을 useGoToCafe 훅으로 분리
